Checking for incorrect decimal separator
There may be a cleaner way of doing this, but this seems to successfully check for the correct decimal separator where parse doesn't.  Perhaps we want to add this to parse_decimals instead of here since we will always want to make sure that the decimal separator is the correct one?